Data extraction complete (Except material details)
