+++ tools/Scene_Extractor/file_id_generator.py
@@ -3,13 +3,14 @@ from unityparser import UnityDocument
 SCENE_FILE_ID = {}
 
 
+
 def generate_file_ids(file):
     file_desc = open(file, "r")
     result = str()
     doc_index = 0
     for lineNumber, line in enumerate(file_desc.readlines()):
         if line.startswith('---'):
-            file_id = line.split('&')[1].split('\n')[0]
+            file_id = line.split('&')[1].split(' ')[0].split('\n')[0]
             SCENE_FILE_ID[file_id] = doc_index
             doc_index += 1
